refactor: 출석률 컴포넌트 분리
